Move facade out of __init__ so that imports are nicer
