+++ src/pydvl/value/least_core/__init__.py
@@ -28,6 +28,8 @@ from pydvl.value.least_core.montecarlo import *
 from pydvl.value.least_core.naive import *
 from pydvl.value.result import ValuationResult
 
+__all__ = ["compute_least_core_values", "LeastCoreMode"]
+
 
 class LeastCoreMode(Enum):
     """Available Least Core algorithms."""
