+++ src/pydvl/value/shapley/owen.py
@@ -12,6 +12,8 @@ from pydvl.utils import MapReduceJob, ParallelConfig, Utility, random_powerset
 from pydvl.value import ValuationResult
 from pydvl.value.stopping import MinUpdates
 
+__all__ = ["OwenAlgorithm", "owen_sampling_shapley"]
+
 
 class OwenAlgorithm(Enum):
     Standard = "standard"
